add runtime layout validation, resolve circular dependency
