+++ pkg/models/types.go
@@ -4,6 +4,13 @@ import (
 	"github.com/canonical/inference-snaps-cli/v2/pkg/engines"
 )
 
+const (
+	capabilityText     string = "text"
+	capabilityVision   string = "vision"
+	capabilityTools    string = "tools"
+	capabilityThinking string = "thinking"
+)
+
 type Manifest struct {
 	ID   string `json:"id" yaml:"id"`
 	Name string `json:"name" yaml:"name"`
@@ -20,4 +27,8 @@ type Manifest struct {
 	Environment []string `json:"environment" yaml:"environment"`
 
 	Layout map[string]engines.Layout `json:"layout,omitempty" yaml:"layout,omitempty"`
+}
+
+func SupportedCapabilities() []string {
+	return []string{capabilityText, capabilityVision, capabilityTools, capabilityThinking}
 }
