+++ pkg/models/validate.go
@@ -10,7 +10,6 @@ import (
 	"slices"
 	"strings"
 
-	"github.com/canonical/inference-snaps-cli/v2/cmd/modelctl/common"
 	"github.com/canonical/inference-snaps-cli/v2/pkg/utils"
 	"go.yaml.in/yaml/v4"
 )
@@ -106,7 +105,7 @@ func (manifest Manifest) validate(expectedModelId string) error {
 		return fmt.Errorf("required field is not set: capabilities")
 	}
 	for _, cap := range manifest.Capabilities {
-		if !slices.Contains(common.SupportedCapabilities(), cap) {
+		if !slices.Contains(SupportedCapabilities(), cap) {
 			return fmt.Errorf("unsupported capability: %q", cap)
 		}
 	}
